== Suggestions for improvement == General:
I think a full module for customer profiles would be in order, while I did create a table to store addresses, I found that it would be more of a headache to try to store and associate
the addresses with a customer based on the simple json posts this api is designed to handle at this point. So instead I approached it from there as more of a guest checkout.
As a guest checkout process, we are ensuring there is no dupe order with in 5 minutes, and we are making sure the products being ordered are in fact products we have, as well as validating
the QTY being passed is a number. To make this possible, I created a line items table, and stored the line items from the sample json. This is now what any new post is compared to.
Further improvements to this would be the following: Inventory tracking so that products are not sold in excess of what is available, or in the cases where the products are made to order,
a notification system to alert the manufacturing department. I would also suggest having the price of the products stored in the line items table, so that if this json post were to also
be in charge of passing pricing, that could be verified as well.

Security:
While the API is currently using _real_escape to escape special characters before inserting any into the database, or in some cases just verifying a field is what it should be, some
regex could potentially be added to look for and remove any common sql injection patterns. Further, a proper user registration system, and storage of their addresses would also increase
security as we could then pull the user's information by record IDs, if those IDs are just numbers, they would be much easier to check for SQL injection.
The api that displays the users order status, while very simple, presents a possible security issue since it is not behind an authentication wall.
I believe the biggest step that could be taken is to add authentication, even if it is not the user authentication, there should be some kind of api key that is passed along so that
this api will only respond to authorized requests. As it stands now, anyone with enough knowledge could post data to this api, and get back the responses.
One other step would be to create a split API, one that is used for the full CRUD operation, and one that only reads. The read one would be used for the un-authenticated requests, and the
SQL statements in that part of the api must select fields intentionally, and not use SELECT *. This way, anywhere the results may be seen, it can only contain fields with less sensitive
information, such as the status of an order, and not all of the order details. While this would not stop a brute for attack for finding the unique view ids, it would prevent the returned data
from being anything more than a simple statement such as 'The order is in processing.' and the like. Further, having the APIs split, and at different URLs, we could effectively hide
the API that has full CRUD access, thus reducing the potential for being targeted by hackers, this is not to replace any other precautions, just an extra that could be worth taking.